backend/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
joblib
nltk
openpyxl
pymongo
python-dotenv
werkzeug
google-auth
google-auth-oauthlib
requests
